requirements.txt
bcrypt==3.2.2
cffi==1.15.1
cryptography==37.0.4
numpy==1.23.1
opencv-python==4.6.0.66
paramiko==2.11.0
Pillow==9.2.0
pycparser==2.21
PyNaCl==1.5.0
pysftp==0.2.9
scipy==1.8.1
six==1.16.0
